Compile the game using renpy8
Everything works as expected.
From the limited testing I did, the only warning/issue that I can
find is that renpy is by default is trying to write/binary a number
of files and that's not compatible with the way flatpak mounts the
/app directory (read-only). These files can eventually be moved to
var/data and be symlink in app, but I'm not sure if that's the
correct way of doing things for renpy.

Other than this, I haven't found any issue with the port and pretty
much everything looked to work (save/load, sound, localization,
videos, etc.)